+++ BetterZeeLog/BetterZeeLog.cs
@@ -14,24 +14,18 @@ using static PluginConfig;
 public sealed class BetterZeeLog : BaseUnityPlugin {
   public const string PluginGUID = "redseiko.valheim.betterzeelog";
   public const string PluginName = "BetterZeeLog";
-  public const string PluginVersion = "1.8.0";
-
-  Harmony _harmony;
+  public const string PluginVersion = "1.9.0";
     
   void Awake() {
     BindConfig(Config);
 
     if (IsModEnabled.Value) {
-      _harmony = Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), harmonyInstanceId: PluginGUID);
+      Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), harmonyInstanceId: PluginGUID);
 
       if (RemoveStackTraceForNonErrorLogType.Value) {
         Application.SetStackTraceLogType(LogType.Log, StackTraceLogType.None);
         Application.SetStackTraceLogType(LogType.Warning, StackTraceLogType.None);
       }
     }
-  }
-
-  void OnDestroy() {
-    _harmony?.UnpatchSelf();
   }
 }