+++ BetterZeeLog/Patches/ZSteamSocketPatch.cs
@@ -11,27 +11,29 @@ using static PluginConfig;
 static class ZSteamSocketPatch {
   [HarmonyTranspiler]
   [HarmonyPatch(nameof(ZSteamSocket.SendQueuedPackages))]
-  static IEnumerable<CodeInstruction> SendQueuedPackagesTranspiler(IEnumerable<CodeInstruction> instructions) {
+  static IEnumerable<CodeInstruction> SendQueuedPackagesTranspiler(
+      IEnumerable<CodeInstruction> instructions, ILGenerator generator) {
     if (RemoveFailedToSendDataLogging.Value) {
-      return new CodeMatcher(instructions)
-          .MatchForward(
-              useEnd: false,
-              new CodeMatch(OpCodes.Ldstr, "Failed to send data "),
-              new CodeMatch(OpCodes.Ldloca_S),
-              new CodeMatch(OpCodes.Constrained),
-              new CodeMatch(OpCodes.Callvirt, AccessTools.Method(typeof(object), nameof(ToString))),
-              new CodeMatch(OpCodes.Call),
+      return new CodeMatcher(instructions, generator)
+          .Start()
+          .MatchStartForward(
+              new CodeMatch(OpCodes.Ldstr, "Failed to send data "))
+          .ThrowIfInvalid($"Could not patch ZSteamSocket.SendQueuedPackages()! (Ldstr-Log)")
+          .ExtractLabels(out List<Label> logLabels)
+          .MatchStartForward(
               new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(ZLog), nameof(ZLog.Log))))
-          .ThrowIfInvalid("Could not patch ZSteamSocket.SendQueuedPackages()! (ZLog.Log)")
+          .ThrowIfInvalid($"Could not patch ZSteamSocket.SendQueuedPackages()! (ZLog-Log)")
           .Advance(offset: 1)
-          .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Pop))
-          .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Nop))
-          .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Nop))
-          .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Nop))
-          .SetInstructionAndAdvance(new CodeInstruction(OpCodes.Nop))
+          .AddLabels(logLabels)
           .InstructionEnumeration();
     }
 
     return instructions;
+  }
+
+  public static CodeMatcher ExtractLabels(this CodeMatcher matcher, out List<Label> labels) {
+    labels = new(matcher.Labels);
+    matcher.Labels.Clear();
+    return matcher;
   }
 }